requirements.txt
discord.py
python-dotenv
d20